refactor!: move scale cast code into its own file
