+++ crates/proof-of-sql/src/sql/mod.rs
@@ -11,4 +11,6 @@ pub mod proof;
 pub mod proof_exprs;
 pub mod proof_gadgets;
 pub mod proof_plans;
+mod scale;
+pub use scale::scale_cast_binary_op;
 pub(crate) mod util;
